Detach ReportFormats into ReportFormatsMixin
