+++ gvm/protocols/gmpv208/__init__.py
@@ -95,6 +95,7 @@ from gvm.protocols.gmpv208.entities.policies import PoliciesMixin
 from gvm.protocols.gmpv208.entities.results import ResultsMixin
 from gvm.protocols.gmpv208.entities.report_formats import (
     ReportFormatType,
+    ReportFormatsMixin,
     get_report_format_id_from_string,
 )
 from gvm.protocols.gmpv208.entities.scan_configs import ScanConfigsMixin
@@ -151,6 +152,7 @@ class Gmp(
     PermissionsMixin,
     PoliciesMixin,
     PortListMixin,
+    ReportFormatsMixin,
     ReportsMixin,
     ResultsMixin,
     TagsMixin,
